// NextMigration/lib/api/types.ts
export interface Studyset {
    id: string;
    title: string;
    course: string;
    global_term_language: string;
    global_definition_language: string;
    created_at: string;
    last_updated: string;
    public_set: boolean;
    displayName: string;
    img_url: string;
    user_id: string;
    folder_id: string;
}
// ...
export interface Visualset {
    id: string;
    title: string;
    course: string;
    created_at: string;
    last_updated: string;
    public_set: boolean;
    user_id: string;
    displayName: string;
    img_url: string;
    folder_id: string;
}
// ...
export interface CreateFolder {
    name: string;
}

export interface UpdateFolder {
    name?: string;
}

export interface Folder {
    id: string;
    name: string;
    owner_id: string;
}

export interface FolderSets {
    studysets: Studyset[];
    visualsets: Visualset[];
}

export interface FullFolder extends Folder {
    sets: FolderSets;
}

export interface FolderListResponse {
    folders: Folder[];
}

export interface SwitchFolderRequest {
    user_id: string;
    set_id: string;
    destinationFolder_id: string;
}